backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
python-multipart==0.0.6

# Pydantic & Settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Cryptography
cryptography==41.0.7

# Redis
redis==5.0.1

# Logging
loguru==0.7.2

# Environment
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Additional utilities
email-validator==2.1.0
python-dateutil==2.8.2
